+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - This is not a V14 update. An update for V14 is still a work in progress.
 ## Bug Fixes:
 - Tactical Mind
+- Deflect Attacks
 ## API Update Notes
 - Added `sockets.syntheticActivityDataRoll` for remotely rolling modified activities, e.g. with changed damage.